+++ cmd/go-qlueless-assembly-api.go
@@ -36,12 +36,20 @@ func main() {
 
 	gorilla := mux.NewRouter()
 
-	gorilla.HandleFunc("/openapi", oai.OpenAPIHandler)
-	gorilla.HandleFunc("/dictionaries", dict.AllDictsHandler)
-	gorilla.HandleFunc("/orders", ord.AllOrdersHandler)
-	gorilla.HandleFunc("/orders/{order_id}", ord.SingleOrderHandler)
-	gorilla.HandleFunc("/batches", bat.AllBatchesHandler)
-	gorilla.HandleFunc("/batches/{batch_id}", bat.SingleBatchHandler)
+	gorilla.HandleFunc("/openapi", oai.OpenAPIHandler).
+		Methods("GET")
+	gorilla.HandleFunc("/dictionaries", dict.AllDictsHandler).
+		Methods("GET")
+	gorilla.HandleFunc("/orders", ord.AllOrdersHandler).
+		Methods("GET")
+	gorilla.HandleFunc("/order", ord.NewOrderHandler).
+		Methods("POST", "OPTIONS")
+	gorilla.HandleFunc("/orders/{order_id}", ord.SingleOrderHandler).
+		Methods("GET")
+	gorilla.HandleFunc("/batches", bat.AllBatchesHandler).
+		Methods("GET")
+	gorilla.HandleFunc("/batches/{batch_id}", bat.SingleBatchHandler).
+		Methods("GET")
 	gorilla.NotFoundHandler = http.HandlerFunc(QluelessNotFoundHandler)
 
 	http.Handle("/", gorilla)